Move VMX-related code from ia32.c to vmx.c
The current ia32.c implements interfaces from both ia32.h and
vmx.h. Move the latter part to vmx.c.

+ Move struct invept_desc from ept.h to vmx.h.
+ Remove redundant #include's from vmx.c, and fix a few whitespace
  errors.
