+++ core/include/vmx.h
@@ -486,7 +486,10 @@ typedef enum encode_t encode_t;
 #define ENCODE_MASK    0x3
 #define ENCODE_SHIFT    13
 
-struct invept_desc;
+struct invept_desc {
+    uint64 eptp;
+    uint64 rsvd;
+};
 
 vmx_result_t ASMCALL asm_invept(uint type, struct invept_desc *desc);
 vmx_result_t ASMCALL asm_vmclear(const paddr_t *addr_in);
